subspace_data_retrieval

Module object_fetcher

Fetching objects stored in the archived history of Subspace Network.

Structs

ObjectFetcher

Object fetcher for the Subspace DSN.

Enums

Error

Object fetching errors.

Functions

max_supported_object_length

The maximum object length the implementation in this module can reliably handle.