Trait subspace_data_retrieval::object_fetcher::ObjectPieceGetter

pub trait ObjectPieceGetter: Debug {
    // Required method
    fn get_piece<'life0, 'async_trait>(
        &'life0 self,
        piece_index: PieceIndex,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Piece>, BoxError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Something that can be used to get decoded pieces by index

Required Methods

fn get_piece<'life0, 'async_trait>( &'life0 self, piece_index: PieceIndex, ) -> Pin<Box<dyn Future<Output = Result<Option<Piece>, BoxError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get piece by index.

Returns Ok(None) for temporary errors: the piece is not found, but immediately retrying this provider might return it. Returns Err(_) for permanent errors: this provider can’t provide the piece at this time, and another provider should be attempted.