subspace_data_retrieval

Module segment_downloading

Fetching segments of the archived history of Subspace Network.

Enums

SegmentDownloadingError

Segment getter errors.

Functions

download_segment

Concurrently downloads the pieces for segment_index, and reconstructs the segment.

download_segment_pieces

Downloads pieces of the segment such that segment can be reconstructed afterward.