Crate subspace_data_retrieval

Fetching data from the archived history of the Subspace Distributed Storage Network.

Modules

object_fetcher

Fetching objects stored in the archived history of Subspace Network.

piece_fetcher

Fetching pieces of the archived history of Subspace Network.

piece_getter

Getting object pieces from the Subspace Distributed Storage Network, or various caches.

segment_downloading

Fetching segments of the archived history of Subspace Network.