subspace_data_retrieval

Module object_fetcher

Fetching objects stored in the archived history of Subspace Network.

Structs

ObjectFetcher

Object fetcher for the Subspace DSN.

Enums

Error

Object fetching errors.

Constants

MAX_BLOCK_LENGTH

Maximum block length for non-Normal extrinsic is 5 MiB. This is a copy of the constant in subspace_runtime_primitives.

Functions

max_supported_object_length

The maximum object length the implementation in this module can reliably handle.

Type Aliases

LastPieceCache

Used to store the last piece downloaded in an object fetcher batch.