Module subspace_data_retrieval::object_fetcher

Fetching objects stored in the archived history of Subspace Network.

Structs

ObjectFetcher

Object fetcher for the Subspace DSN.

Enums

Error

Object fetching errors.

PieceGetterError

Piece getter errors.

SegmentGetterError

Segment getter errors.

Traits

ObjectPieceGetter

Something that can be used to get decoded pieces by index

Type Aliases

BoxError

A type-erased error