subspace_data_retrieval

Module segment_downloading

Fetching segments of the archived history of Subspace Network.

Enums

SegmentDownloadingError

Segment getter errors.

Constants

SEGMENT_DOWNLOAD_RETRIES

The number of times we try to download a segment before giving up. This is a suggested default, callers can supply their own value if needed.

SEGMENT_DOWNLOAD_RETRY_DELAY

The amount of time we wait between segment download retries. This is a suggested default, callers can supply their own value if needed.

Functions

download_segment

Concurrently downloads the pieces for segment_index, and reconstructs the segment.

download_segment_pieces

Downloads pieces of a segment so that segment can be reconstructed afterward. Repeatedly attempts to download pieces until the required number of pieces is reached.